requirements.txt
streamlit==1.38.0
pdfplumber==0.11.0
google-generativeai==0.5.3
python-dotenv==1.0.1
scikit-learn==1.5.0
